+++ CHANGELOG.md
@@ -15,6 +15,8 @@ Wherever possible, a reference to an issue in the project issue tracker should b
   [#15](https://gitlab.data.bas.ac.uk/MAGIC/ops-data-store/-/issues/15)
 * `config show` CLI command
   [#18](https://gitlab.data.bas.ac.uk/MAGIC/ops-data-store/-/issues/18)
+* documentation and tests for the config class
+  [#20](https://gitlab.data.bas.ac.uk/MAGIC/ops-data-store/-/issues/20)
 
 ### Changed